Fixed CaptureAward message
ProtoBufjs throws an illegal wire type error if the fields are not explicitly marked as packed. According to the Protocol Buffers docs, a wire type of 2, which, according to the error message, is used for packed fields among other things. The second example in example.js parses CatchPokemonResponse s after this change.